Added support for gRPC
Changes:
- Added a gRPC service for BERT functions including answer, discriminate, and predict.
- Added a gRPC service for NER functions including analyze.
- Added a gRPC server that listens on port 1976.
- Added subcommands to bert_server and ner-server CLI for the gRPC client.
- Renamed `bert_server run` to `bert_server server`.
- Renamed `ner-server run` to `ner-server server`.
- Refactored by extracting functions and moving handler functions to different compilation units.

Closes #35.
