+++ go.mod
@@ -5,6 +5,7 @@ go 1.14
 require (
 	github.com/dgraph-io/badger/v2 v2.0.3
 	github.com/dustin/go-humanize v1.0.0
+	github.com/golang/protobuf v1.4.2
 	github.com/gosuri/uilive v0.0.4 // indirect
 	github.com/gosuri/uiprogress v0.0.1
 	github.com/lithammer/fuzzysearch v1.1.0
@@ -15,5 +16,10 @@ require (
 	github.com/stretchr/testify v1.4.0
 	github.com/urfave/cli v1.22.4
 	golang.org/x/exp v0.0.0-20191227195350-da58074b4299
+	golang.org/x/net v0.0.0-20200625001655-4c5254603344 // indirect
+	golang.org/x/sys v0.0.0-20200625212154-ddb9806d33ae // indirect
+	golang.org/x/text v0.3.3 // indirect
 	gonum.org/v1/gonum v0.6.2
+	google.golang.org/grpc v1.27.0
+	google.golang.org/protobuf v1.25.0
 )
