+++ pkg/nlp/sequencelabeler/grpcapi/sequencelabeler.proto
@@ -1,0 +1,34 @@
+syntax = "proto3";
+package grpcapi;
+
+// The SequenceLabeler service definition.
+service SequenceLabeler {
+  
+  // Sends a request to /analyze.
+  rpc Analyze(AnalyzeRequest) returns (AnalyzeReply) {}
+}
+
+// The analyze request message containing the tokens for the sequence labeler analysis.
+message AnalyzeRequest {
+  string text              = 1;
+  bool   mergeEntities     = 2;
+  bool   filterNotEntities = 3;
+
+  // Took is the number of milliseconds it took the server to execute the request.
+	int64  took              = 4;
+}
+
+// The analyze response message containing the text to analyze with options.
+message Token {
+  string text  = 1;
+	int32  start = 2;
+	int32  end   = 3;
+	string label = 4;
+}
+
+message AnalyzeReply {
+  repeated Token tokens = 1;
+
+  // Took is the number of milliseconds it took the server to execute the request.
+	int64          took   = 2;
+}
