+++ pkg/nlp/sequencelabeler/server.go
@@ -4,22 +4,22 @@
 
 package sequencelabeler
 
-import (
-	"bytes"
-	"encoding/json"
-	"github.com/nlpodyssey/spago/pkg/webui/ner"
-	"log"
-	"net/http"
-	"time"
+//go:generate protoc --go_out=Mgrpc/service_config/service_config.proto=/internal/proto/grpc_service_config:.  --go-grpc_out=Mgrpc/service_config/service_config.proto=/internal/proto/grpc_service_config:. --go_opt=paths=source_relative --go-grpc_opt=paths=source_relative grpcapi/sequencelabeler.proto
 
-	"github.com/nlpodyssey/spago/pkg/ml/ag"
-	"github.com/nlpodyssey/spago/pkg/ml/nn"
-	"github.com/nlpodyssey/spago/pkg/nlp/tokenizers/basetokenizer"
-	"github.com/nlpodyssey/spago/pkg/utils/httphandlers"
+import (
+	"net/http"
+
+	"github.com/nlpodyssey/spago/pkg/nlp/sequencelabeler/grpcapi"
+	"github.com/nlpodyssey/spago/pkg/utils/grpcutils"
+	"github.com/nlpodyssey/spago/pkg/utils/httputils"
+	"github.com/nlpodyssey/spago/pkg/webui/ner"
 )
 
 type Server struct {
 	model *Model
+
+	// UnimplementedSequenceLabelerServer must be embedded to have forward compatible implementations for gRPC.
+	grpcapi.UnimplementedSequenceLabelerServer
 }
 
 func NewServer(model *Model) *Server {
@@ -28,111 +28,14 @@ func NewServer(model *Model) *Server {
 	}
 }
 
-func (s *Server) Start(address, tlsCert, tlsKey string, tlsDisable bool) {
-	r := http.NewServeMux()
-	r.HandleFunc("/ner-ui", ner.Handler)
-	r.HandleFunc("/analyze", s.analyze)
+func (s *Server) Start(address, grpcAddress, tlsCert, tlsKey string, tlsDisable bool) {
+	mux := http.NewServeMux()
+	mux.HandleFunc("/ner-ui", ner.Handler)
+	mux.HandleFunc("/analyze", s.analyze)
 
-	if tlsDisable {
-		log.Fatal(http.ListenAndServe(address,
-			httphandlers.RecoveryHandler(httphandlers.PrintRecoveryStack(true))(r)))
-	} else {
-		log.Fatal(http.ListenAndServeTLS(address, tlsCert, tlsKey,
-			httphandlers.RecoveryHandler(httphandlers.PrintRecoveryStack(true))(r)))
-	}
-}
+	go httputils.RunHTTPServer(address, tlsDisable, tlsCert, tlsKey, mux)
 
-type OptionsType struct {
-	MergeEntities     bool `json:"mergeEntities"`     // default false
-	FilterNotEntities bool `json:"filterNotEntities"` // default false
-}
-
-type Body struct {
-	Options OptionsType `json:"options"`
-	Text    string      `json:"text"`
-}
-
-func (s *Server) analyze(w http.ResponseWriter, req *http.Request) {
-	w.Header().Set("Access-Control-Allow-Origin", "*") // that's intended for testing purposes only
-	w.Header().Set("Content-Type", "application/json")
-
-	var body Body
-	err := json.NewDecoder(req.Body).Decode(&body)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusBadRequest)
-		return
-	}
-
-	analysis, took := s.process(body.Text, body.Options.MergeEntities)
-	if body.Options.FilterNotEntities {
-		analysis = filterNotEntities(analysis)
-	}
-	result := prepareResponse(analysis, took)
-
-	_, pretty := req.URL.Query()["pretty"]
-	response, err := result.Dump(pretty)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-
-	_, err = w.Write(response)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-}
-
-func (s *Server) process(text string, merge bool) ([]TokenLabel, time.Duration) {
-	start := time.Now()
-	g := ag.NewGraph()
-	defer g.Clear()
-	proc := s.model.NewProc(g).(*Processor)
-	proc.SetMode(nn.Inference)
-	tokenized := basetokenizer.New().Tokenize(text)
-	predicted := proc.Predict(tokenized)
-	if merge {
-		predicted = mergeEntities(predicted)
-	}
-	return predicted, time.Since(start)
-}
-
-func prepareResponse(tokens []TokenLabel, took time.Duration) *Response {
-	newTokens := make([]Token, len(tokens))
-	for i, token := range tokens {
-		newTokens[i] = Token{
-			Text:  token.String,
-			Start: token.Offsets.Start,
-			End:   token.Offsets.End,
-			Label: token.Label,
-		}
-	}
-	return &Response{Tokens: newTokens, Took: took.Milliseconds()}
-}
-
-type Response struct {
-	Tokens []Token `json:"tokens"`
-	// Took is the number of milliseconds it took the server to execute the request.
-	Took int64 `json:"took"`
-}
-
-type Token struct {
-	Text  string `json:"text"`
-	Start int    `json:"start"`
-	End   int    `json:"end"`
-	Label string `json:"label"`
-}
-
-func (r *Response) Dump(pretty bool) ([]byte, error) {
-	buf := bytes.NewBufferString("")
-	enc := json.NewEncoder(buf)
-	if pretty {
-		enc.SetIndent("", "    ")
-	}
-	enc.SetEscapeHTML(true)
-	err := enc.Encode(r)
-	if err != nil {
-		return nil, err
-	}
-	return buf.Bytes(), nil
+	grpcServer := grpcutils.NewGRPCServer(tlsDisable, tlsCert, tlsKey)
+	grpcapi.RegisterSequenceLabelerServer(grpcServer, s)
+	grpcutils.RunGRPCServer(grpcAddress, grpcServer)
 }
